Get rid of ParameterUpdateException
_update_param used to raise a ParameterUpdateException if one of the new
values did not match at least one of the current value objects. However,
this was dropped to better support the case where the parameters are being
extended along some dimension to fill the parameter space. An exception could
be raised if a new value object contains a dimension that is not used in the
current value objects for the parameter. However, it seems like it could be
expensive to check this case, especially when a project is extending parameters.
For now, no exceptions are raised by this method.